Extract same code from bundle deb & rpm into new bundle_linux_distro file
